+++ test/manual/test_trtllm_fp8_kv_kernel.py
@@ -6,7 +6,9 @@ import unittest
 
 import torch
 
-from sglang.srt.layers.attention.trtllm_fp8_kv_kernel import fused_fp8_set_kv_buffer
+from sglang.srt.layers.attention.triton_ops.trtllm_fp8_kv_kernel import (
+    fused_fp8_set_kv_buffer,
+)
 from sglang.test.test_utils import CustomTestCase
